+++ layers/Schema/packages/migrate-tags-wp/migrate/contractimplementations/cms-apis/load.php
@@ -1,4 +1,3 @@
 <?php
 
-require_once 'cms-objectpropertyresolver.php';
 require_once 'cms-routingstate-hooks.php';
