--- layers/Schema/packages/migrate-tags/composer.json
@@ -1,28 +1,0 @@
-{
-    "name": "pop-schema/migrate-tags",
-    "type": "library",
-    "description": "(Temporary package) Migrate code to package https://github.com/pop-schema/tags. Target package description: Adds support for tags",
-    "keywords": ["PoP", "tags"],
-    "homepage": "https://github.com/pop-schema/migrate-tags",
-    "license": "GPL-2.0-or-later",
-    "authors": [
-        {
-            "name": "Leonardo Losoviz",
-            "email": "leo@getpop.org",
-            "homepage": "https://getpop.org"
-        }
-    ],
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "require": {
-        "php": "^8.0"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.8-dev"
-        }
-    },
-    "config": {
-        "sort-packages": true
-    }
-}
