--- layers/Schema/packages/migrate-tags/migrate/contracts/cms-apis/cms-functionapi-interface.php
@@ -1,8 +1,0 @@
-<?php
-namespace PoPSchema\Tags;
-
-use PoPSchema\Taxonomies\TypeAPIs\TaxonomyTypeAPIInterface;
-
-interface FunctionAPI extends TaxonomyTypeAPIInterface
-{
-}
