--- layers/Schema/packages/migrate-tags/migrate/contracts/cms-apis/cms-objectpropertyresolver-base.php
@@ -1,10 +1,0 @@
-<?php
-namespace PoPSchema\Tags;
-
-abstract class ObjectPropertyResolver_Base implements ObjectPropertyResolver
-{
-    public function __construct()
-    {
-        ObjectPropertyResolverFactory::setInstance($this);
-    }
-}
