--- layers/Schema/packages/migrate-tags/migrate/contracts/cms-apis/cms-objectpropertyresolver-interface.php
@@ -1,11 +1,0 @@
-<?php
-namespace PoPSchema\Tags;
-
-interface ObjectPropertyResolver
-{
-    public function getTagName($tag);
-    public function getTagSlug($tag);
-    public function getTagDescription($tag);
-    public function getTagCount($tag);
-    public function getTagID($tag);
-}
