--- layers/Schema/packages/migrate-tags/migrate/contracts/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'cms-apis/load.php';
