+++ layers/Schema/packages/post-tags/src/ComponentContracts/PostTagAPISatisfiedContractTrait.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace PoPSchema\PostTags\ComponentContracts;
 
 use PoPSchema\PostTags\TypeResolvers\PostTagTypeResolver;
+use PoPSchema\Tags\TypeAPIs\TagTypeAPIInterface;
 
 trait PostTagAPISatisfiedContractTrait
 {
-    protected function getTypeAPI(): \PoPSchema\Tags\FunctionAPI
+    protected function getTypeAPI(): TagTypeAPIInterface
     {
         $cmstagsapi = \PoPSchema\PostTags\FunctionAPIFactory::getInstance();
         return $cmstagsapi;
