+++ layers/Schema/packages/tags/src/ComponentContracts/TagAPIRequestedContractTrait.php
@@ -4,13 +4,16 @@ declare(strict_types=1);
 
 namespace PoPSchema\Tags\ComponentContracts;
 
+use PoPSchema\Tags\Facades\TagTypeAPIFacade;
+use PoPSchema\Tags\TypeAPIs\TagTypeAPIInterface;
+
 trait TagAPIRequestedContractTrait
 {
-    abstract protected function getTypeAPI(): \PoPSchema\Tags\FunctionAPI;
+    abstract protected function getTypeAPI(): TagTypeAPIInterface;
     abstract protected function getTypeResolverClass(): string;
-    protected function getObjectPropertyAPI(): \PoPSchema\Tags\ObjectPropertyResolver
+    protected function getObjectPropertyAPI(): TagTypeAPIInterface
     {
-        $cmstagsresolver = \PoPSchema\Tags\ObjectPropertyResolverFactory::getInstance();
+        $cmstagsresolver = TagTypeAPIFacade::getInstance();
         return $cmstagsresolver;
     }
 }
